style: corrige lint C408 y round() para ruff 0.16 (CI)
El CI (ruff sin pin, instala la ultima) usa reglas de default mas amplias que
mi ruff local viejo. Tres hallazgos:
- test_montecarlo.py: dos `dict(...)` -> literales `{...}` (C408).
- simulation.py (sembrar_estado): `int(round(x))` -> `round(x)`, que ya devuelve
  int (regla nueva; venia de PR #4, no era del Montecarlo).

ruff 0.16 limpio sobre src+tests; 165 tests en verde.

+++ src/astrobiosim/simulation.py
@@ -97,7 +97,7 @@ def sembrar_estado(
     if patron == "uniforme":
         estado[rng.random(shape) < fraccion_activa] = ACTIVA
     elif patron == "cluster":
-        lado = max(1, min(int(round((fraccion_activa * m * n) ** 0.5)), m, n))
+        lado = max(1, min(round((fraccion_activa * m * n) ** 0.5), m, n))
         i0, j0 = (m - lado) // 2, (n - lado) // 2
         estado[i0 : i0 + lado, j0 : j0 + lado] = ACTIVA
     else:
